Bugfix 1: Call read_char() after checking for an invalid string.
As is, the function returns Token::Type::Illegal for all strings, because it is calling read_char()
before checking if the last character read was '\0'. When the next character to be read is '\0', the
subsequent if check would consider that as an invalid string and return an illegal token.

Bugfix 2: Subtract 1 from pos - orig_pos when returning a valid string.

As is, the function returns a token with a literal containing the ending
\".